Move mock data file
